Address warpctrl foundation review feedback
